Involved inducing sanitation at a higher level for the original API Call. Now the home page has the logic which turns the API call into a cleaner object handed down to CardsContainer. CardsContainer can now receive those objects from either the saved page or the home page, which both receive their data from different sources so this sanitization was necessary.
TODO: Further root out any remnants of logic in CardsContainer and try
and elevate it where necessary.

TODO: Add a prop to to CardsContainer so it can tell where the data is
coming from and toss in a delete or save button. Data from saved should
be accompanied by a delete button so you can remove it from your
respository. Data from the home directory should come with a saved
button to add it to your repository.